// src/lib/hetzner/types.gen.ts
export type CreateZoneRrsetData = {
  body: {
    /**
     * Name of the [RRSet](#tag/zone-rrsets).
     *
     * The name must be in lower case, and must not end with a dot or the [Zone](#tag/zones)
     * name. Names containing non-ASCII characters must be transcribed to
     * [Punycode](https://wikipedia.org/wiki/Punycode) representation with ACE prefix, e.g.
     * `xn--4bi` (✉️).
     *
     * For the [Zone](#tag/zones) apex, use `@`.
     *
     */
    name: string;
    /**
     * Type of the [RRSet](#tag/zone-rrsets).
     *
     */
    type:
      | 'A'
      | 'AAAA'
      | 'CAA'
      | 'CNAME'
      | 'DS'
      | 'HINFO'
      | 'HTTPS'
      | 'MX'
      | 'NS'
      | 'PTR'
      | 'RP'
      | 'SOA'
      | 'SRV'
      | 'SVCB'
      | 'TLSA'
      | 'TXT';
    /**
     * Time To Live (TTL) of the [RRSet](#tag/zone-rrsets).
     *
     * Must be in between 60s and 2147483647s.
     *
     * If not set, the [Zone's](#tag/zones) Default TTL is used.
     *
     */
    ttl?: number | null;
    /**
     * Records of the [RRSet](#tag/zone-rrsets).
     *
     * Must not be empty and must only contain distinct record values.
     * The order of records returned in responses is not guaranteed to be consistent.
     *
     */
    records: Array<{
      /**
       * Value of the record.
       *
       * For details about accepted values, see the [DNS record types documentation](https://docs.hetzner.com/networking/dns/record-types/overview/).
       *
       */
      value: string;
      /**
       * Comment of the record.
       */
      comment?: string;
    }>;
    /**
     * User-defined labels (`key/value` pairs) for the Resource.
     * For more information, see "[Labels](#description/labels)".
     *
     */
    labels?: {
      [key: string]: string;
    };
  };
  path: {
    /**
     * ID or Name of the [Zone](#tag/zones).
     */
    id_or_name: string;
  };
  query?: never;
  url: '/zones/{id_or_name}/rrsets';
};
// ...
export type SetZoneRrsetRecordsData = {
  body: {
    /**
     * Records to set in the [RRSet](#tag/zone-rrsets).
     *
     * Must not be empty and must only contain distinct record values.
     *
     */
    records: Array<{
      /**
       * Value of the record.
       *
       * For details about accepted values, see the [DNS record types documentation](https://docs.hetzner.com/networking/dns/record-types/overview/).
       *
       */
      value: string;
      /**
       * Comment of the record.
       */
      comment?: string;
    }>;
  };
  path: {
    /**
     * ID or Name of the [Zone](#tag/zones).
     */
    id_or_name: string;
    /**
     * Name of the [RRSet](#tag/zone-rrsets).
     *
     * The name must be in lower case, and must not end with a dot or the [Zone](#tag/zones)
     * name. Names containing non-ASCII characters must be transcribed to
     * [Punycode](https://wikipedia.org/wiki/Punycode) representation with ACE prefix, e.g.
     * `xn--4bi` (✉️).
     *
     * For the [Zone](#tag/zones) apex, use `@`.
     *
     */
    rr_name: string;
    /**
     * Type of the [RRSet](#tag/zone-rrsets).
     *
     */
    rr_type:
      | 'A'
      | 'AAAA'
      | 'CAA'
      | 'CNAME'
      | 'DS'
      | 'HINFO'
      | 'HTTPS'
      | 'MX'
      | 'NS'
      | 'PTR'
      | 'RP'
      | 'SOA'
      | 'SRV'
      | 'SVCB'
      | 'TLSA'
      | 'TXT';
  };
  query?: never;
  url: '/zones/{id_or_name}/rrsets/{rr_name}/{rr_type}/actions/set_records';
};